ported game from c to c++
