+++ Makefile
@@ -1,4 +1,4 @@
 
 
 default:
-	clang main.c -o main -Llib -lraylib -Iinclude -ldl -lGL -lm -lpthread -lrt -lX11
+	clang++ main.cpp -o main -Llib -lraylib -Iinclude -ldl -lGL -lm -lpthread -lrt -lX11
