--- endConditions.c
@@ -1,5 +1,0 @@
-#pragma once
-
-int neverEnd() {
-    return 0;
-}
