+++ endConditions.cpp
@@ -1,0 +1,7 @@
+#include "gameLoop.cpp"
+
+#pragma once
+
+bool neverEnd(GameLogic* state) {
+    return false;
+}
