--- readOnlyStrings.c
@@ -1,9 +1,0 @@
-# pragma once
-
-#define rStringMake(string) {string, sizeof(string)}
-
-typedef struct {
-    char* stringPointer;
-    int lenght;
-} rString;
-
