+++ wordGenerators.cpp
@@ -1,0 +1,11 @@
+#include <iostream>
+
+using namespace std;
+
+#pragma once
+
+string penguinString = "penguin";
+
+string penguin() {
+    return penguinString;
+}
